Reorganise current attachment models
